+++ src/test/java/com/moyorak/api/team/service/TeamRestaurantFacadeTest.java
@@ -1,7 +1,6 @@
 package com.moyorak.api.team.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.mockito.BDDMockito.given;
 
@@ -19,7 +18,6 @@ import com.moyorak.api.review.service.ReviewService;
 import com.moyorak.api.team.domain.SortOption;
 import com.moyorak.api.team.domain.TeamRestaurant;
 import com.moyorak.api.team.domain.TeamRestaurantFixture;
-import com.moyorak.api.team.domain.TeamRestaurantNotFoundException;
 import com.moyorak.api.team.domain.TeamRestaurantSummaries;
 import com.moyorak.api.team.dto.TeamRestaurantListRequest;
 import com.moyorak.api.team.dto.TeamRestaurantListRequestFixture;
@@ -28,10 +26,8 @@ import com.moyorak.api.team.dto.TeamRestaurantResponse;
 import com.moyorak.api.team.dto.TeamRestaurantSummary;
 import com.moyorak.api.team.dto.TeamRestaurantSummaryFixture;
 import com.moyorak.api.team.repository.TeamRestaurantRepository;
-import com.moyorak.config.exception.BusinessException;
 import com.moyorak.global.domain.ListResponse;
 import java.util.List;
-import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -167,77 +163,6 @@ class TeamRestaurantFacadeTest {
                         it.assertThat(response.servingTime()).isEqualTo("5분이내");
                         it.assertThat(response.waitingTime()).isEqualTo("5분이내");
                     });
-        }
-
-        @Test
-        @DisplayName("팀 맛집이 존재하지 않으면 예외가 발생합니다.")
-        void throwsWhenTeamRestaurantNotFound() {
-            // given
-            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
-                    .willReturn(Optional.empty());
-
-            // when & then
-            assertThatThrownBy(
-                            () ->
-                                    teamRestaurantFacade.getTeamRestaurant(
-                                            userId, teamId, teamRestaurantId))
-                    .isInstanceOf(TeamRestaurantNotFoundException.class);
-        }
-
-        @Test
-        @DisplayName("식당이 null이면 예외가 발생합니다.")
-        void throwsWhenRestaurantIsNull() {
-            // given
-            final TeamRestaurant teamRestaurant =
-                    TeamRestaurantFixture.fixture(
-                            teamRestaurantId, "맛있네요", 4.5, 5, 5, 5.5, 5, true, teamId, null);
-
-            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
-                    .willReturn(Optional.of(teamRestaurant));
-
-            // when & then
-            assertThatThrownBy(
-                            () ->
-                                    teamRestaurantFacade.getTeamRestaurant(
-                                            userId, teamId, teamRestaurantId))
-                    .isInstanceOf(BusinessException.class)
-                    .hasMessage("연결된 식당 정보가 존재하지 않습니다.");
-        }
-
-        @Test
-        @DisplayName("use_yn이 'N'인 경우 조회되지 않아 예외가 발생합니다.")
-        void throwsWhenUseIsFalse() {
-            // given
-            final Restaurant restaurant =
-                    RestaurantFixture.fixture(
-                            "http://place.map.kakao.com/000000",
-                            "쓰지 않는 식당",
-                            "서울시 어디구",
-                            "서울로 456",
-                            RestaurantCategory.KOREAN,
-                            127.0,
-                            37.0);
-            final TeamRestaurant disabled =
-                    TeamRestaurantFixture.fixture(
-                            teamRestaurantId,
-                            "쓰지 않는 식당",
-                            4.0,
-                            5,
-                            5,
-                            5.0,
-                            5,
-                            false,
-                            teamId,
-                            restaurant);
-            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
-                    .willReturn(Optional.empty());
-
-            // when & then
-            assertThatThrownBy(
-                            () ->
-                                    teamRestaurantFacade.getTeamRestaurant(
-                                            userId, teamId, teamRestaurantId))
-                    .isInstanceOf(TeamRestaurantNotFoundException.class);
         }
     }
 }
