+++ src/test/java/com/moyorak/api/team/service/TeamRestaurantServiceTest.java
@@ -18,6 +18,7 @@ import com.moyorak.api.team.domain.Team;
 import com.moyorak.api.team.domain.TeamFixture;
 import com.moyorak.api.team.domain.TeamRestaurant;
 import com.moyorak.api.team.domain.TeamRestaurantFixture;
+import com.moyorak.api.team.domain.TeamRestaurantNotFoundException;
 import com.moyorak.api.team.domain.TeamRestaurantSearch;
 import com.moyorak.api.team.domain.TeamRestaurantSearchFixture;
 import com.moyorak.api.team.domain.TeamUser;
@@ -47,6 +48,85 @@ class TeamRestaurantServiceTest {
     @Mock private TeamUserRepository teamUserRepository;
     @Mock private RestaurantRepository restaurantRepository;
     @Mock private TeamRestaurantSearchRepository teamRestaurantSearchRepository;
+
+    @Nested
+    @DisplayName("팀 맛집 조회 시,")
+    class GetTeamRestaurant {
+        private final Long teamId = 1L;
+        private final Long teamRestaurantId = 1L;
+        private final Long userId = 1L;
+
+        @Test
+        @DisplayName("팀 맛집이 존재하지 않으면 예외가 발생합니다.")
+        void throwsWhenTeamRestaurantNotFound() {
+            // given
+            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
+                    .willReturn(Optional.empty());
+
+            // when & then
+            assertThatThrownBy(
+                            () ->
+                                    teamRestaurantService.getValidatedTeamRestaurant(
+                                            teamId, teamRestaurantId))
+                    .isInstanceOf(TeamRestaurantNotFoundException.class);
+        }
+
+        @Test
+        @DisplayName("식당이 null이면 예외가 발생합니다.")
+        void throwsWhenRestaurantIsNull() {
+            // given
+            final TeamRestaurant teamRestaurant =
+                    TeamRestaurantFixture.fixture(
+                            teamRestaurantId, "맛있네요", 4.5, 5, 5, 5.5, 5, true, teamId, null);
+
+            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
+                    .willReturn(Optional.of(teamRestaurant));
+
+            // when & then
+            assertThatThrownBy(
+                            () ->
+                                    teamRestaurantService.getValidatedTeamRestaurant(
+                                            teamId, teamRestaurantId))
+                    .isInstanceOf(BusinessException.class)
+                    .hasMessage("연결된 식당 정보가 존재하지 않습니다.");
+        }
+
+        @Test
+        @DisplayName("use_yn이 'N'인 경우 조회되지 않아 예외가 발생합니다.")
+        void throwsWhenUseIsFalse() {
+            // given
+            final Restaurant restaurant =
+                    RestaurantFixture.fixture(
+                            "http://place.map.kakao.com/000000",
+                            "쓰지 않는 식당",
+                            "서울시 어디구",
+                            "서울로 456",
+                            RestaurantCategory.KOREAN,
+                            127.0,
+                            37.0);
+            final TeamRestaurant disabled =
+                    TeamRestaurantFixture.fixture(
+                            teamRestaurantId,
+                            "쓰지 않는 식당",
+                            4.0,
+                            5,
+                            5,
+                            5.0,
+                            5,
+                            false,
+                            teamId,
+                            restaurant);
+            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
+                    .willReturn(Optional.empty());
+
+            // when & then
+            assertThatThrownBy(
+                            () ->
+                                    teamRestaurantService.getValidatedTeamRestaurant(
+                                            teamId, teamRestaurantId))
+                    .isInstanceOf(TeamRestaurantNotFoundException.class);
+        }
+    }
 
     @Nested
     @DisplayName("팀 맛집 저장 시")
